Separated NOTE funcs / structs to their own files
