+++ tests/note_tests.c
@@ -1,0 +1,45 @@
+#include "catch.hpp"
+#include "../code/merc.h"
+#include "../code/note.h"
+
+
+// TODO: These functions rely on database calls. Need to find a good c++ mocking framework for databases.
+
+// TODO: write tests
+//int count_spool (CHAR_DATA *ch, int type);
+
+// TODO: write tests
+//void do_unread (CHAR_DATA *ch, char *arg);
+
+// TODO: write tests
+//void do_note (CHAR_DATA *ch,char *argument);
+
+// TODO: write tests
+//void do_news (CHAR_DATA *ch,char *argument);
+
+// TODO: write tests
+//void do_changes (CHAR_DATA *ch,char *argument);
+
+// TODO: write tests
+//void append_note (NOTE_DATA *pnote);
+
+// TODO: write tests
+//bool is_note_to (CHAR_DATA *ch, char *sender, char *to_list);
+
+// TODO: write tests
+//void note_attach (CHAR_DATA *ch, int type);
+
+// TODO: write tests
+//bool hide_note (CHAR_DATA *ch, MYSQL_ROW row);
+
+// TODO: write tests
+//void update_read (CHAR_DATA *ch, long stamp, int type);
+
+// TODO: write tests
+//void parse_note (CHAR_DATA *ch, char *argument, int type);
+
+// TODO: write tests
+//NOTE_DATA *new_note(void);
+
+// TODO: write tests
+//void free_note(NOTE_DATA *note);
